adjust Trailing sapce
